fix: delete view all plaltforms link btn rename folder
