+++ wealth-sync/src/app/page.tsx
@@ -1,7 +1,7 @@
 import Footer from "@/components/Footer";
 import IntroductionSection from "@/components/IntroductionSection";
 import FeaturesSection from "@/components/FeaturesSection";
-import IntegrationSection from "@/components/IntegrationSection";
+import IntegrationSection from "@/components/IntegrationPlatformsSection/IntegrationSection";
 import PricingSection from "@/components/PricingSection";
 import ChartPrievewSection from "@/components/ChartPrievewSection";
 import ActionSection from "@/components/ActionSection";
