+++ wealth-sync/src/components/IntegrationPlatformsSection/IntegrationPlatformCard.tsx
