+++ wealth-sync/src/components/IntegrationPlatformsSection/IntegrationPlatformsGrid.tsx
